增加权限，修改头像

+++ backend/controllers/PediaUserGroupController.php
@@ -17,6 +17,7 @@ use yii\filters\VerbFilter;
 /**
  * Team:没有蛀牙
  * Coding by:孙一冉 1711297，20190712
+ * Coding by:解亚兰 1711431，20190713
  * This is the controller of pedia-user-group
  */
 class PediaUserGroupController extends Controller
@@ -81,8 +82,8 @@ class PediaUserGroupController extends Controller
         $pid = PediaUserGroup::find()->where(['gid' => $gid])->asArray()->one()['pid'];
         $edit = PediaUserPerm::find()->where(['pid' => $pid])->asArray()->one()['allowedcreword'];
         if ($edit != 1) {
-            echo "<script>alert('新人不允许新增词条')</script>";
-            return $this->goHome();
+            ?><script>alert("只有管理员可以创建用户组");history.back();</script><?php
+            exit("0");
         }
         $this->layout='backcon';
         $model = new PediaUserGroup();
@@ -105,6 +106,13 @@ class PediaUserGroupController extends Controller
      */
     public function actionUpdate($id)
     {
+        $gid = PediaUserMember::find()->where(['loginname' => Yii::$app->user->identity->username])->asArray()->one()['gid'];
+        $pid = PediaUserGroup::find()->where(['gid' => $gid])->asArray()->one()['pid'];
+        $edit = PediaUserPerm::find()->where(['pid' => $pid])->asArray()->one()['allowedcreword'];
+        if ($edit != 1) {
+            ?><script>alert("只有管理员可以更改用户组");history.back();</script><?php
+            exit("0");
+        }
         $this->layout='backcon';
         $model = $this->findModel($id);
 
@@ -126,6 +134,13 @@ class PediaUserGroupController extends Controller
      */
     public function actionDelete($id)
     {
+        $gid = PediaUserMember::find()->where(['loginname' => Yii::$app->user->identity->username])->asArray()->one()['gid'];
+        $pid = PediaUserGroup::find()->where(['gid' => $gid])->asArray()->one()['pid'];
+        $edit = PediaUserPerm::find()->where(['pid' => $pid])->asArray()->one()['allowedcreword'];
+        if ($edit != 1) {
+            ?><script>alert("只有管理员可以删除用户组");history.back();</script><?php
+            exit("0");
+        }
         $this->layout='backcon';
         $this->findModel($id)->delete();
